Move Entity/Handle SDK APIs to core
