+++ java/arcs/core/storage/api/BUILD
@@ -7,10 +7,13 @@ licenses(["notice"])
 
 package(default_visibility = ["//visibility:public"])
 
+SDK_INTERFACES = ["Entity.kt", "Handle.kt"]
+
 arcs_kt_library(
     name = "api",
-    srcs = glob(["*.kt"]),
+    srcs = glob(["*.kt"], exclude = SDK_INTERFACES),
     deps = [
+        ":api-common",
         "//java/arcs/core/common",
         "//java/arcs/core/crdt",
         "//java/arcs/core/data",
@@ -22,6 +25,15 @@ arcs_kt_library(
         "//java/arcs/core/storage/referencemode",
         "//java/arcs/core/type",
         "//java/arcs/core/util",
+        "//third_party/kotlin/kotlinx_coroutines",
+    ],
+)
+
+arcs_kt_library(
+    name = "api-common",
+    srcs = SDK_INTERFACES,
+    deps = [
+        "//java/arcs/core/data:rawentity",
         "//third_party/kotlin/kotlinx_coroutines",
     ],
 )
